# Hackaton15/SV77217548/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hackaton15 API",
    "version": "1.0.0",
    "description": "API de autenticacion con sesion y JWT"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Bearer <token>"
    },
    "sessionAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "sid",
      "description": "Cookie de sesion"
    },
    "refreshAuth": {
      "type": "apiKey",
      "in": "cookie",
      "name": "refreshToken",
      "description": "Cookie de refresh token"
    }
  },
  "paths": {
...
    "/session/login": {
      "post": {
        "tags": [
          "Session"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/session/logout": {
      "post": {
        "tags": [
          "Session"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/me": {
      "get": {
        "tags": [
          "Session"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  }
}